Add summary preference
This commit allows the user to specify a 'summarize' preference
setting. When set, gptcli will summarize the full conversation thread
w/ GPT 3.5 Turbo when a user adds to it with a new prompt. The summary
is then sent to the GPT 4.0 Turbo model rather than the fully
conversation thread. When set this preference is trading reduced
precision of responses from GPT 4.0 Turbo for lower costs.