Rename persisted threads
Currently threads are persisted by generating a filename based on the
CRC of the name of the thread. However, if there are different threads
with the same name this will cause a collision. This commit appends
the creation timestamp of a thread in order to avoid
collisions. Additionally, it will rename any files using the old
naming scheme to the new one.