Add Yesterday and subcommand aliases
When showing current threads, create an alias for 'Yesterday' to make
it visually easier for users to find a thread when they have many of
them. Additionally, allow users to save on typing with command
aliases. e.g. instead of 'archive' accept 'a' if there are no other
subcommands that start with an 'a'.