Introduce proxy ui
ncurses is single threaded, so ensure that we only ever update the
ncurses ui from a single goroutine. in order to facilitate this, this
commit introduces the notion of a "proxy ui" where approval requests
can be queued in the invoking goroutine, and then serviced later by
the goroutine responsible for the ui with results communicated via channel.